chore: Success, Loading 뷰 파일 분리
