+++ feature/spot/src/main/java/com/acon/acon/feature/spot/screen/spotlist/composable/SpotListScreen.kt
@@ -91,8 +91,6 @@ import kotlinx.collections.immutable.toImmutableSet
 import kotlinx.coroutines.launch
 import kotlin.math.absoluteValue
 
-private const val MAX_GUEST_AVAILABLE_COUNT = 5
-
 @Composable
 internal fun SpotListScreen(
     state: SpotListUiStateV2,
@@ -249,172 +247,6 @@ internal fun SpotListScreen(
             )
             is SpotListUiStateV2.LoadFailed -> {
                 // TODO("Error")
-            }
-        }
-    }
-}
-
-@Composable
-private fun SpotListSuccessView(
-    pagerState: PagerState,
-    state: SpotListUiStateV2.Success,
-    userType: UserType,
-    onSpotClick: (SpotV2) -> Unit,
-    onGuestItemClick: () -> Unit,
-    onTryFindWay: (SpotV2) -> Unit,
-    itemHeightPx: Float,
-    modifier: Modifier = Modifier,
-) {
-
-    val context = LocalContext.current
-
-    VerticalPager(
-        state = pagerState,
-        contentPadding = PaddingValues(
-            horizontal = 16.dp,
-            vertical = (itemHeightPx * .18f).toDp()
-        ),
-        modifier = modifier,
-        flingBehavior = PagerDefaults.flingBehavior(
-            state = pagerState,
-            pagerSnapDistance = PagerSnapDistance.atMost(4),
-            decayAnimationSpec = exponentialDecay(
-                frictionMultiplier = 1.2f
-            )
-        ),
-        horizontalAlignment = Alignment.CenterHorizontally,
-        snapPosition = SnapPosition.Center,
-        pageSize = PageSize.Fixed((itemHeightPx).toDp())
-    ) { page ->
-        val spot = state.spotList[page]
-        var spotFogColor by remember {
-            mutableStateOf(Color.Transparent)
-        }
-
-        Column(
-            horizontalAlignment = Alignment.CenterHorizontally,
-            modifier = Modifier
-                .hazeSource(LocalHazeState.current)
-                .padding(vertical = 6.dp)
-                .fillMaxWidth(.91f)
-                .graphicsLayer {
-                    val pageOffset =
-                        pagerState.getOffsetDistanceInPages(page).absoluteValue
-                    val ratio = lerp(
-                        start = 1.1f,
-                        stop = 0.9f,
-                        fraction = pageOffset.coerceIn(0f, 1f)
-                    )
-                    scaleX = ratio
-                    scaleY = ratio
-                }
-        ) {
-            if (page == 0) {
-                Text(
-                    text = "최고의 선택.",
-                    style = AconTheme.typography.Title2,
-                    fontWeight = FontWeight.SemiBold,
-                    color = AconTheme.color.White,
-                    modifier = Modifier
-                        .padding(bottom = 6.dp)
-                        .fogBackground(
-                            glowColor = AconTheme.color.White,
-                            glowRadius = 100f
-                        )
-                        .zIndex(2f)
-                )
-            }
-            if (page >= MAX_GUEST_AVAILABLE_COUNT && userType == UserType.GUEST) {
-                SpotGuestItem(
-                    modifier = Modifier
-                        .fillMaxSize()
-                        .clip(RoundedCornerShape(20.dp))
-                        .background(
-                            shape = RoundedCornerShape(20.dp),
-                            color = AconTheme.color.GlassBlackDefault
-                        )
-                        .clickable {
-                            onGuestItemClick()
-                        }
-                        .fogBackground(
-                            glowColor = AconTheme.color.White,
-                        )
-                )
-            } else {
-                SpotItem(
-                    spot = spot,
-                    onItemClick = onSpotClick,
-                    onFindWayButtonClick = onTryFindWay,
-                    modifier = Modifier
-                        .fillMaxSize()
-                        .fogBackground(
-                            glowColor = spotFogColor,
-                            glowAlpha = 1f,
-                        )
-                        .zIndex(1f)
-                )
-            }
-        }
-
-        LaunchedEffect(Unit) {
-            spotFogColor = spot.image.getOverlayColor(context)
-        }
-    }
-}
-
-@Composable
-private fun SpotListLoadingView(
-    itemHeightPx: Float,
-    modifier: Modifier = Modifier,
-) {
-    Column(
-        modifier = modifier,
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        SkeletonItem(
-            modifier = Modifier
-                .width(120.dp)
-                .height(28.dp),
-            shape = RoundedCornerShape(8.dp),
-        )
-        repeat(10) {
-            Spacer(modifier = Modifier.height(14.dp))
-            Column(
-                modifier = Modifier
-                    .height(itemHeightPx.toDp())
-                    .fillMaxWidth()
-                    .skeleton(
-                        shape = RoundedCornerShape(20.dp)
-                    )
-                    .padding(20.dp)
-            ) {
-                Row(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .height(26.dp)
-                ) {
-                    SkeletonItem(
-                        modifier = Modifier
-                            .fillMaxHeight()
-                            .weight(3.5f),
-                        shape = RoundedCornerShape(8.dp),
-                    )
-                    SkeletonItem(
-                        modifier = Modifier
-                            .padding(start = 10.dp)
-                            .fillMaxHeight()
-                            .weight(1f),
-                        shape = RoundedCornerShape(8.dp),
-                    )
-                }
-                Spacer(modifier = Modifier.weight(1f))
-                SkeletonItem(
-                    modifier = Modifier
-                        .height(36.dp)
-                        .width(140.dp)
-                        .align(Alignment.End),
-                    shape = RoundedCornerShape(8.dp),
-                )
             }
         }
     }
